pyproject.toml
[project]
name = "pdf-parser-benchmark"
version = "0.1.0"
description = "Benchmark suite for PDF-to-text parsers with built-in test dataset"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.11.7",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "rich>=14.1.0",
    # Evaluation & AI
    "openai>=1.99.0",
    "google-genai>=1.46.0",
    "mistralai>=1.9.0",
    "requests>=2.32.0",
    "levenshtein>=0.27.1",
    "nltk>=3.9.1",
    # Used for Unicode to LaTeX conversion in formula rendering
    "pylatexenc>=2.10",
    # Visualization
    "gradio>=5.41.0",
]

[project.optional-dependencies]
# Synthetic PDF generation - for creating test documents
synth_pdf = [
    "faker>=37.11.0",
    "duckdb>=1.4.1",
]

# PDF Parsers - install individually as needed
pymupdf = [
    "pymupdf4llm>=0.0.17",
]

# All extras combined
all = [
    "pdf-parser-benchmark[synth_pdf,pymupdf]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_benchmark", "parsers"]

[tool.hatch.build.targets.wheel.force-include]
"data" = "pdf_benchmark/data"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/data",
]
